fix: move  out of root so it can be pub
