+++ validator_derive/src/tokens/contains.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::Contains;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn contains_tokens(
     crate_name: &CrateName,
