+++ validator_derive/src/tokens/does_not_contain.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::DoesNotContain;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn does_not_contain_tokens(
     crate_name: &CrateName,
