+++ validator_derive/src/tokens/ip.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::Ip;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn ip_tokens(
     crate_name: &CrateName,
