+++ validator_derive/src/tokens/length.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::Length;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn length_tokens(
     crate_name: &CrateName,
