+++ validator_derive/src/tokens/must_match.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::MustMatch;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn must_match_tokens(
     crate_name: &CrateName,
