+++ validator_derive/src/tokens/non_control_character.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::NonControlCharacter;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn non_control_char_tokens(
     crate_name: &CrateName,
