+++ validator_derive/src/tokens/regex.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::Regex;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn regex_tokens(
     crate_name: &CrateName,
