+++ validator_derive/src/tokens/url.rs
@@ -1,8 +1,7 @@
 use quote::quote;
 
 use crate::types::Url;
-use crate::utils::{quote_code, quote_message};
-use crate::CrateName;
+use crate::utils::{quote_code, quote_message, CrateName};
 
 pub fn url_tokens(
     crate_name: &CrateName,
